feat: Add find_or_create_game method to QueueService
The `find_or_create_game` method has been added to the `QueueService` class. This method allows users to find an available opponent for a specific game type and create a new game if an opponent is found. If no opponent is available, the user is added to the queue.

This commit message follows the established convention of using a prefix to indicate the type of change (`feat` for a new feature) and provides a clear and concise description of the code changes.